python-shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "redis>=5.0.0",
    "dotenv>=0.9.9",
    "pydantic_settings>=2.12.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
